requirements.txt
pandas
numpy
scikit-learn
matplotlib
umap-learn
scipy
